create hak akses setiap role(otorisasi)

+++ app/category/controller.js
@@ -12,6 +12,14 @@ async function index(req, res, next) {
 
 async function store(req, res, next) {
 	try {
+		let policy = policyFor(req.user);
+		if (!policy.can('create', 'Category')) {
+			return res.json({
+				error: 1,
+				message: 'Anda tidak memiliki akses untuk membuat kategori'
+			});
+		}
+
 		let payload = req.body;
 		let category = new Category(payload);
 		await category.save();
@@ -30,6 +38,14 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
 	try {
+		let policy = policyFor(req.user);
+		if (!policy.can('update', 'Category')) {
+			return res.json({
+				error: 1,
+				message: 'Anda tidak memiliki akses untuk mengupdate kategori'
+			});
+		}
+
 		let payload = req.body;
 		let category = await Category.findOneAndUpdate({ _id: req.params.id }, payload, {
 			new: true,
@@ -50,6 +66,13 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
 	try {
+		let policy = policyFor(req.user);
+		if (!policy.can('delete', 'Category')) {
+			return res.json({
+				error: 1,
+				message: 'Anda tidak memiliki akses untuk menghapus kategori'
+			});
+		}
 		let deleted = await Category.findOneAndDelete({ _id: req.params.id });
 		return res.json(deleted);
 	} catch (err) {
